pipecat-shopping-agent/requirements.txt
# Pipecat framework with only needed services
pipecat-ai[daily,google]

# Additional dependencies for enhanced functionality  
python-dotenv
loguru